update number_formant and modify price type in database

+++ src/helpers/Functions.php
@@ -11,7 +11,7 @@ function url($endPoint)
 // Format price
 function formatNumber($num)
 {
-    return (float)number_format(
+    return number_format(
         $num,
         2,
         ",",
